# src/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: API RESTful - Trisog
  version: 1.0.0
  description: >-
    This project involves the development of a **RESTful API using Node.js and
    TypeScript**, with a focus on integrating **MongoDB** for database
    management. The API is designed to manage **touristic experiences**,
    allowing users to interact with data related to destinations, activities,
    bookings, and reviews.


    The API will enable users to perform essential operations such as creating,
    retrieving, updating, and deleting information about destinations and
    activities. Additionally, it will support booking management and user
    reviews for these experiences. Security and data privacy are key priorities,
    with robust authentication and authorization mechanisms implemented using
    **Firebase Authentication**.


    **Firebase Storage** is utilized for managing and storing user-uploaded
    images associated with experiences. **Docker** and **Docker Compose** are
    employed to ensure consistent and scalable deployment across different
    environments. **Jest** is used for unit and integration testing, ensuring
    the reliability and quality of the API.


    **Prisma** is used as the ORM tool to facilitate interaction with MongoDB,
    providing a robust and type-safe interface for database queries and data
    manipulation. This allows for efficient and secure data handling throughout
    the application.


...


...


...
  license:
    name: MIT License
...
components:
  securitySchemes:
    jwt:
      type: apiKey
      in: header
      name: Authorization
      description: To authenticate, pass the JWT token in the format **Bearer _token_**.
  schemas:
...
    ExperienceList:
      type: array
      items:
        $ref: '#/components/schemas/Experience'
...
    PaginatedExperiences:
      type: object
      required:
        - page
        - limit
        - total_pages
        - total_experiences
        - experiences
      properties:
        page:
          type: integer
          description: The current page number
          example: 1
        limit:
          type: integer
          description: Number of items per page
          example: 10
        total_pages:
          type: integer
          description: Total number of pages
          example: 5
        total_experiences:
          type: integer
          description: Total number of experiences
          example: 50
        experiences:
          type: array
          items:
            $ref: '#/components/schemas/ExperienceList'
          description: List of experiences for the current page
...
    ExperienceUpsert:
      type: object
      required:
        - title
        - city
        - image
        - video
        - gallery
        - map_link
        - start_date
        - end_date
        - duration
        - is_activity
        - max_people
        - min_age
        - over_view
        - include
        - exclude
        - default_price
        - categories_id
        - plans_id
        - destination_id
      properties:
        title:
          type: string
          description: Title of the experience
          example: City Tour
        city:
          type: string
          description: City where the experience takes place
          example: New York
        image:
          type: string
          description: URL to an image representing the experience
          example: https://example.com/image.jpg
        video:
          type: string
          description: URL to a video related to the experience
          example: https://example.com/video.mp4
        gallery:
          type: string
          description: URL to a gallery related to the experience
          example: https://example.com/gallery
        map_link:
          type: string
          description: URL to the map location of the experience
          example: https://example.com/map
        start_date:
          type: string
          format: date-time
          description: Start date of the availability period YYYY-MM-DD
          example: '2024-03-24'
        end_date:
          type: string
          format: date-time
          description: End date of the availability period YYYY-MM-DD
          example: '2024-03-24'
        duration:
          type: number
          description: Duration of experience in minutes
          example: 60
        is_activity:
          type: boolean
          description: Indicates if the experience is an activity
          example: false
        max_people:
          type: integer
          description: Maximum number of people allowed for the experience
          example: 30
        min_age:
          type: integer
          description: Minimum age required for participation in the experience
          example: 18
        over_view:
          type: string
          description: Overview of the experience
          example: A comprehensive tour of the city with stops at major landmarks.
        include:
          type: array
          items:
            type: string
          description: List of items included in the experience
          example:
            - Lunch
            - Guide
        exclude:
          type: array
          items:
            type: string
          description: List of items excluded from the experience
          example:
            - Transportation
        default_price:
          type: number
          description: The default price applied when no custom price is specified
          example: 4.5
        custom_prices:
          type: array
          description: An array of custom prices for specific dates
          items:
            $ref: '#/components/schemas/CustomPrice'
        categories_id:
          type: array
          items:
            type: string
            description: List of category IDs associated with the experience
          description: List of category IDs associated with the experience
          example:
            - 60b8d295f1c0d2b0f1b2c3d4
            - 60b8d295f1c0d2b0f1b2c3d5
        plans_id:
          type: array
          items:
            type: string
          description: List of plan IDs associated with the experience
          example:
            - 60b8d295f1c0d2b0f1b2c3d6
            - 60b8d295f1c0d2b0f1b2c3d7
        destination_id:
          type: string
          description: ID of the destination associated with the experience
          example: 63f47adf2d7e6b04e4f978c7
...
    DeleteSuccess:
      type: object
      properties:
        msg:
          type: string
          description: Delete success message
          example: Successfully deleted
    BadRequest:
      type: object
      properties:
        msg:
          type: string
          description: Invalid request. Some parameters may be missing or invalid.
          example: Some parameters may be missing or invalid
    Unauthorized:
      type: object
      properties:
        msg:
          type: string
          description: Unauthorized. The JWT token was not provided or is invalid.
          example: JWT token not provided or invalid
    Forbidden:
      type: object
      properties:
        msg:
          type: string
          description: >-
            Forbidden. The user does not have permission to access the requested
            resource.
          example: You do not have permission to access the requested resource
    NotFound:
      type: object
      properties:
        msg:
          type: string
          description: No data found. The requested resource does not exist
          example: No data found
    InternalServerError:
      type: object
      properties:
        msg:
          type: string
          description: Internal error when handling data in the service
          example: Internal Server Error
    ReviewList:
      type: array
      items:
        $ref: '#/components/schemas/Review'
...
paths:
...
  /experiences:
    get:
      summary: View all experiences
      description: Returns information about all experiences
      tags:
        - Experience
      operationId: get_all_experiences
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: string
            default: '1'
        - name: limit
          in: query
          description: Number of experiences per page
          required: false
          schema:
            type: string
            default: '10'
        - name: title
          in: query
          description: Title to search for in experiences
          required: false
          schema:
            type: string
            example: Tour
        - name: price
          in: query
          description: Maximum price to filter experiences by
          required: false
          schema:
            type: string
            example: '150'
        - name: categoriesId
          in: query
          description: Comma-separated list of category IDs to filter by
          required: false
          schema:
            type: string
            example: 66ce29934244142ada6b021e,66ce29934244142ada6b021f
        - name: destinationsId
          in: query
          description: Comma-separated list of destination IDs to filter by
          required: false
          schema:
            type: string
            example: 76ce29934244142ada6b021e,76ce29934244142ada6b021f
        - name: rating
          in: query
          description: Minimum rating to filter experiences by
          required: false
          schema:
            type: string
            example: '4'
        - name: date
          in: query
          description: Date to filter experiences by
          required: false
          schema:
            type: string
            format: date
            example: '2024-03-24'
        - name: guests
          in: query
          description: Number of guests to filter by
          required: false
          schema:
            type: string
            example: '2'
        - name: isActivity
          in: query
          description: Filter by whether the experience is an activity (true/false)
          required: false
          schema:
            type: string
            enum:
              - 'true'
              - 'false'
            example: 'false'
        - name: sortBy
          in: query
          description: Field to sort experiences by
          required: false
          schema:
            type: string
            example: title
        - name: order
          in: query
          description: Sort order (ascending or descending)
          required: false
          schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedExperiences'
        '400':
          description: BadRequest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '500':
          description: InternalServerError
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
    post:
      summary: Create a new experience
      description: Creates a new experience based on the data provided in the request body.
      tags:
        - Experience
      operationId: create_experience
      security:
        - jwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExperienceUpsert'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Experience'
        '400':
          description: BadRequest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '404':
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: InternalServerError
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /experiences/user/favorites:
    get:
      summary: See all favorited experiences
      description: Returns information about all favorited experiences
      tags:
        - Experience
      operationId: get_all_favorited_experiences
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: string
            default: '1'
        - name: limit
          in: query
          description: Number of experiences per page
          required: false
          schema:
            type: string
            default: '10'
        - name: title
          in: query
          description: Title to search for in experiences
          required: false
          schema:
            type: string
            example: Tour
        - name: price
          in: query
          description: Maximum price to filter experiences by
          required: false
          schema:
            type: string
            example: '150'
        - name: categoriesId
          in: query
          description: Comma-separated list of category IDs to filter by
          required: false
          schema:
            type: string
            example: 66ce29934244142ada6b021e,66ce29934244142ada6b021f
        - name: destinationsId
          in: query
          description: Comma-separated list of destination IDs to filter by
          required: false
          schema:
            type: string
            example: 76ce29934244142ada6b021e,76ce29934244142ada6b021f
        - name: rating
          in: query
          description: Minimum rating to filter experiences by
          required: false
          schema:
            type: string
            example: '4'
        - name: date
          in: query
          description: Date to filter experiences by
          required: false
          schema:
            type: string
            format: date
            example: '2024-03-24'
        - name: guests
          in: query
          description: Number of guests to filter by
          required: false
          schema:
            type: string
            example: '2'
        - name: isActivity
          in: query
          description: Filter by whether the experience is an activity (true/false)
          required: false
          schema:
            type: string
            enum:
              - 'true'
              - 'false'
            example: 'false'
        - name: sortBy
          in: query
          description: Field to sort experiences by
          required: false
          schema:
            type: string
            example: title
        - name: order
          in: query
          description: Sort order (ascending or descending)
          required: false
          schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedExperiences'
        '400':
          description: BadRequest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '500':
          description: InternalServerError
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /experiences/{experience_id}:
    parameters:
      - name: experience_id
        in: path
        required: true
        description: ID of the experience to be deleted
        schema:
          type: string
    get:
      summary: View experience by id
      description: Returns experience information by id
      tags:
        - Experience
      operationId: get_experience_by_id
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Experience'
        '400':
          description: BadRequest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '404':
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: InternalServerError
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
    put:
      summary: Update experience
      description: Updates the information of the desired experience
      tags:
        - Experience
      operationId: update_experience
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExperienceUpsert'
      security:
        - jwt: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Experience'
        '400':
          description: BadRequest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '404':
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: InternalServerError
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
    delete:
      summary: Remove experience
      description: Deletes a experience based on the ID provided
      tags:
        - Experience
      operationId: delete_experience
      security:
        - jwt: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteSuccess'
        '400':
          description: BadRequest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '404':
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: InternalServerError
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
...
  /reviews/experience/{experience_id}:
    parameters:
      - name: experience_id
        in: path
        required: true
        description: ID of the experience that will fetch the reviews
        schema:
          type: string
    get:
      summary: View review by experience
      description: Returns review information by experience
      tags:
        - Review
      operationId: get_review_by_experience
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewList'
        '400':
          description: BadRequest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '404':
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: InternalServerError
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
...